requirements.txt
fastapi
uvicorn
prometheus-client
sqlalchemy
python-dotenv
pydantic